backend/django_app_structure.py

"""
ARCHITECTURE: /backend/src/api/v1/
Purpose: Production-grade Django implementation for Equipment Analytics.
"""

from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework import status, permissions, authentication
import pandas as pd
import io

class EquipmentUploadView(APIView):
    """
    ARCHITECTURE: /backend/src/api/v1/uploads/
    Purpose: Secured endpoint for CSV ingestion and analytics.
    """
    authentication_classes = [authentication.BasicAuthentication]
    permission_classes = [permissions.IsAuthenticated]

    def post(self, request):
        file = request.FILES.get('file')
        if not file:
            return Response({"error": "No file uploaded"}, status=status.HTTP_400_BAD_REQUEST)
        
        try:
            # Production-grade parsing with Pandas
            df = pd.read_csv(file)
            
            # Column Validation
            required_cols = ['Equipment Name', 'Type', 'Flowrate', 'Pressure', 'Temperature']
            if not all(col in df.columns for col in required_cols):
                return Response(
                    {"error": f"Invalid CSV schema. Required: {required_cols}"}, 
                    status=status.HTTP_422_UNPROCESSABLE_ENTITY
                )

            # Analytics Calculation
            summary = {
                "total_count": len(df),
                "avg_flowrate": float(df['Flowrate'].mean()),
                "avg_pressure": float(df['Pressure'].mean()),
                "avg_temperature": float(df['Temperature'].mean()),
                "type_distribution": df['Type'].value_counts().to_dict()
            }
            
            # Logic to maintain only last 5 datasets would go here (Database persistence)
            
            return Response({
                "id": "ds_" + str(hash(file.name)),
                "filename": file.name,
                "summary": summary,
                "data": df.to_dict(orient='records')
            }, status=status.HTTP_201_CREATED)
            
        except Exception as e:
            return Response({"error": f"Processing failed: {str(e)}"}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)

class DatasetHistoryView(APIView):
    """
    ARCHITECTURE: /backend/src/api/v1/equipment/
    Purpose: Returns metadata of the last 5 successful uploads.
    """
    authentication_classes = [authentication.BasicAuthentication]
    permission_classes = [permissions.IsAuthenticated]

    def get(self, request):
        # Implementation would query SQLite for the latest 5 entries in DatasetHistory table
        return Response([], status=status.HTTP_200_OK)
