backend/requirements.txt
# ChemEquip Visualizer - Backend Dependencies
# Django REST Framework API for Equipment Analytics

# Core Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0

# CORS Support for Web Frontend
django-cors-headers>=4.0.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0  # Excel file support (.xlsx, .xls)

# PDF Report Generation
reportlab>=4.0.0

# Database (SQLite is included with Python, but for production PostgreSQL)
# psycopg2-binary>=2.9.0  # Uncomment for PostgreSQL support

# Security & Authentication
# django-allauth>=0.54.0  # Uncomment for advanced auth

# Development Tools (optional)
# django-debug-toolbar>=4.0.0  # Uncomment for debugging
